# (chore) Bloom gateway: Improve fingerprint partitioning in client