**What this PR does / why we need it**:

This PR re-implements the fingerprint partitioning on the bloom gateway client using `slices.BinarySearchFunc`.

Benchmark results from `sort.Search` (old) vs. `slices.BinarySearchFunc` (new):

```
$ benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/bloomgateway
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                                   │   old.txt    │              new.txt               │
                                   │    sec/op    │   sec/op     vs base               │
PartitionFingerprintsByAddresses-8   18.75µ ± 70%   16.94µ ± 2%  -9.68% (p=0.000 n=10)

                                   │   old.txt    │             new.txt              │
                                   │     B/op     │     B/op      vs base            │
PartitionFingerprintsByAddresses-8   18.69Ki ± 0%   18.69Ki ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                                   │  old.txt   │            new.txt             │
                                   │ allocs/op  │ allocs/op   vs base            │
PartitionFingerprintsByAddresses-8   8.000 ± 0%   8.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```